Amusing Joke

+++ Codeforces/Amusing Joke.cpp
@@ -3,8 +3,8 @@
 #define HASH 13
 using namespace std;
 
-int hsh (string s) {
-    int ans = 0;
+long long hsh (string s) {
+    long long ans = 0;
     
     for (char c : s)
         ans += c - 'a' * pow(HASH, c - 'a');
@@ -15,8 +15,8 @@ int hsh (string s) {
 int main() {
     string a, b, c;
     cin >> a >> b >> c;
-    int prev = hsh(a) + hsh(b);
-    int cur = hsh(c);
+    long long prev = hsh(a) + hsh(b);
+    long long cur = hsh(c);
     
     if (prev == cur) printf("YES");
     else printf("NO");
